Moves /import-message to /messages/broadcast
