+++ app/routes/import-message.js
@@ -8,12 +8,12 @@ const router = express.Router();
 const outboundMessageConfig = require('../../config/lib/middleware/import-message/message-outbound');
 
 // Middleware
-const paramsMiddleware = require('../../lib/middleware/import-message/params');
-const getBroadcastMiddleware = require('../../lib/middleware/import-message/broadcast-get');
-const parseBroadcastMiddleware = require('../../lib/middleware/import-message/parse-broadcast');
+const paramsMiddleware = require('../../lib/middleware/messages/broadcast/params');
+const getBroadcastMiddleware = require('../../lib/middleware/messages/broadcast/broadcast-get');
+const parseBroadcastMiddleware = require('../../lib/middleware/messages/broadcast/parse-broadcast');
 const getConvoMiddleware = require('../../lib/middleware/conversation-get');
 const createConvoMiddleware = require('../../lib/middleware/conversation-create');
-const updateConvoMiddleware = require('../../lib/middleware/import-message/conversation-update');
+const updateConvoMiddleware = require('../../lib/middleware/messages/broadcast/conversation-update');
 const loadOutboundMessageMiddleware = require('../../lib/middleware/message-outbound-load');
 const createOutboundMessageMiddleware = require('../../lib/middleware/message-outbound-create');
