+++ lib/middleware/messages/broadcast/broadcast-get.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const logger = require('../../logger');
-const helpers = require('../../helpers');
-const contentful = require('../../contentful');
-const UnprocessibleEntityError = require('../../../app/exceptions/UnprocessibleEntityError');
+const logger = require('../../../logger');
+const helpers = require('../../../helpers');
+const contentful = require('../../../contentful');
+const UnprocessibleEntityError = require('../../../../app/exceptions/UnprocessibleEntityError');
 
 module.exports = function getBroadcast() {
   return (req, res, next) => {
