+++ lib/middleware/messages/broadcast/conversation-update.js
@@ -1,9 +1,8 @@
 'use strict';
 
-const logger = require('../../logger');
-
-const helpers = require('../../helpers');
-const gambitCampaigns = require('../../gambit-campaigns');
+const logger = require('../../../logger');
+const helpers = require('../../../helpers');
+const gambitCampaigns = require('../../../gambit-campaigns');
 
 module.exports = function updateConversation() {
   return (req, res, next) => {
