+++ lib/middleware/messages/broadcast/params.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const logger = require('../../logger');
-const helpers = require('../../helpers');
-const UnprocessibleEntityError = require('../../../app/exceptions/UnprocessibleEntityError');
+const logger = require('../../../logger');
+const helpers = require('../../../helpers');
+const UnprocessibleEntityError = require('../../../../app/exceptions/UnprocessibleEntityError');
 
 module.exports = function params() {
   return (req, res, next) => {
