+++ lib/middleware/messages/broadcast/parse-broadcast.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const helpers = require('../../helpers');
-const contentful = require('../../contentful');
+const helpers = require('../../../helpers');
+const contentful = require('../../../contentful');
 
 module.exports = function parseBroadcast() {
   return (req, res, next) => {
